Refactor config into default file + env provision
Rename config.js to config/default.js
Extract env loader into config/index.js
